REFACTOR mongodb connection logic
